requirements.txt
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-cors==0.1
django-cors-headers==3.1.0
django-countries==5.4
django-guardian==2.0.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-twilio==0.10.0
djangorestframework==3.10.2
djangorestframework-guardian==0.3.0
djangorestframework-jwt==1.11.0
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
Pillow==6.1.0
PyJWT==1.7.1
PySocks==1.7.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
twilio==6.29.3
urllib3==1.25.3
whitenoise==4.1.3
